+++ src/app/user/[name]/page.tsx
@@ -1,10 +1,12 @@
-import ProfileContainer from "@/components/profile/profileContainer";
 import * as S from "./page.css";
-import RankInfoContainer from "@/components/ranking/rankInfoContainer";
-import GameRecordContainer from "@/components/record/gameRecordContainer";
 import { Suspense } from "react";
-import ProfilePreviewLayoutContainer from "@/components/profile/profilePreviewLayoutContainer";
-import RankInfoLayoutContainer from "@/components/ranking/rankInfoLayoutContainer";
+
+import {
+  GameRecordContainer,
+  RankInfoContainer,
+  RankInfoLayoutContainer,
+} from "@/components/record";
+import { ProfileContainer, ProfilePreviewLayoutContainer } from "@/components/profile";
 
 const UserRecord = () => {
   return (
