+++ src/components/common/ItemListBox/index.css.ts
@@ -1,6 +1,6 @@
 import { theme } from "@/style/base/theme/index.css";
 import { Row } from "@/style/ui/Flex";
-import { createVar, style } from "@vanilla-extract/css";
+import { style } from "@vanilla-extract/css";
 
 export const Container = style(
   Row({
