+++ src/components/common/index.ts
@@ -1,0 +1,8 @@
+export { default as ChampionInfoImg } from "./championInfoImg";
+export { default as Header } from "./header";
+export { default as ItemListBox } from "./ItemListBox";
+export { default as KDAavgText } from "./KDAavgText";
+export { default as KDAText } from "./KDAText";
+export { default as RunespellGridBox } from "./runespellGridBox";
+export { default as UserChampionDetailInfoSection } from "./userChampionDetailInfoSection";
+export { default as Provider } from "./provider";
