+++ src/components/ranking/index.ts
@@ -1,5 +1,3 @@
-export { default as RankInfoContainer } from "./rankInfoContainer";
-export { default as RankInfoLayoutContainer } from "./rankInfoLayoutContainer";
 export { default as RankInfoSection } from "./rankInfoSection";
 export { default as SchoolRankHeaderContainer } from "./schoolRankHeaderContainer";
 export { default as SchoolRankUserContainer } from "./schoolRankUserContainer";
