+++ src/components/record/gameRecordContainer/gameRecordHeaderLink/index.tsx
@@ -1,5 +1,5 @@
-import { assignInlineVars } from "@vanilla-extract/dynamic";
 import * as S from "./index.css";
+import { assignInlineVars } from "@vanilla-extract/dynamic";
 import { GameRecordHeaderLinkProperties } from "@/types/components/GameRecordHeaderLinkProperties.type";
 import { theme } from "@/style/base/theme/index.css";
