+++ src/components/record/index.ts
@@ -1,2 +1,4 @@
 export { default as GameRecordContainer } from "./gameRecordContainer";
 export { default as PlayedChampionListContainer } from "./playedChampionListContainer";
+export { default as RankInfoContainer } from "../record/rankInfoContainer";
+export { default as RankInfoLayoutContainer } from "../record/rankInfoLayoutContainer";
