+++ src/components/record/rankInfoContainer/index.css.ts
