+++ src/components/record/rankInfoLayoutContainer/index.css.ts
