+++ src/components/record/rankInfoLayoutContainer/index.tsx
