+++ src/types/components/AdditionalInfoModalProperties.type.ts
@@ -1,4 +1,3 @@
 export interface AdditionalModalInfoProperties {
   animationState: boolean;
-  closeModal: () => void;
 }
